(lwidsplugin: r37187)

+++ lwidsplugin/LWICRC.cpp
@@ -1,6 +1,7 @@
 
 #include "LWICRC.h"
 #include <time.h>
+#include <stdint.h>
 
 LWICRC* LWICRC::_instance = NULL;
 
@@ -53,32 +54,32 @@ LWICRC::InitTable()
 {
     assert (_key != 0);
     
-	// for all possible byte values
-	for (unsigned i = 0; i < 256; ++i)
-	{
-		unsigned long reg = i << 24;
-		// for all bits in a byte
-		for (int j = 0; j < 8; ++j)
-		{
-			bool topBit = (reg & 0x80000000) != 0;
-			reg <<= 1;
-			if (topBit)
+    // for all possible byte values
+    for (int i = 0; i < 256; ++i)
+    {
+        uint32_t reg = i << 24;
+        // for all bits in a byte
+        for (int j = 0; j < 8; ++j)
+        {
+            bool topBit = (reg & 0x80000000) != 0;
+            reg <<= 1;
+            if (topBit)
             {
-				reg ^= _key;
+                reg ^= _key;
             }
-		}
-		_table [i] = reg;
-	}
+        }
+        _table [i] = reg;
+    }
 }
 
 unsigned long
 LWICRC::CalculateCRC(char* pByteArray, int length) const
 {
-    unsigned long reg = 0;
+    uint32_t reg = 0;
     
     for (int i = 0; i < length; i++)
     {
-        unsigned top = reg >> 24;
+        uint32_t top = reg >> 24;
         top ^= *(pByteArray+i);
         reg = (reg << 8) ^ _table [top];
     }
